+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-10521 Changed always showing switches on the members page
 * VFS-8257 Added closing modal on transition
 * VFS-9257 Fixed ignored eslint issues
 * VFS-11006 Changed EGI auth icon